Extract creators to resolve proliferation of constructor dependencies
