+++ devtools_m2/cypress/plugins/index.js
@@ -23,7 +23,7 @@ module.exports = (on, config) => {
   config.defaultCommandTimeout = 10000 // Default: 4000
   // config.execTimeout = 60000 // Default: 60000
   // config.taskTimeout = 60000 // Default: 60000
-  // config.pageLoadTimeout = 60000 // Default: 60000
+  config.pageLoadTimeout = 100000 // Default: 60000
   config.requestTimeout = 10000 // Default: 5000
   config.responseTimeout = 10000 // Default: 3000
 }
