+++ devtools_m2/cypress/support/commands.js
@@ -23,3 +23,10 @@
 //
 // -- This is will overwrite an existing command --
 // Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
+
+afterEach(function() {
+  // https://github.com/cypress-io/cypress/issues/216
+  // https://github.com/cypress-io/cypress/issues/686
+  cy.log('Reset web interface to avoid bleed-over')
+  cy.visit('/lib/web/blank.html', { failOnStatusCode: false })
+})
